+++ src/main/resources/db/migration/V006__cql_rename.sql
@@ -1,0 +1,2 @@
+UPDATE samply.query SET format = 'CQL_DATA' WHERE format = 'CQL';
+UPDATE samply.query SET format = 'CQL' WHERE format = 'CQL_MEASURE_REPORT';